# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Document:
    properties:
      chunk_count:
        type: integer
      create_time:
        type: string
      created_by:
        type: string
      error_message:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      kb_id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/model.DocumentStatus'
      update_time:
        type: string
    type: object
  model.DocumentStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-comments:
      DocumentStatusCompleted: 已完成
      DocumentStatusFailed: 失败
      DocumentStatusPending: 待处理
      DocumentStatusProcessing: 处理中
    x-enum-descriptions:
    - 待处理
    - 处理中
    - 已完成
    - 失败
    x-enum-varnames:
    - DocumentStatusPending
    - DocumentStatusProcessing
    - DocumentStatusCompleted
    - DocumentStatusFailed
  model.KnowledgeBase:
    properties:
      collection_name:
        type: string
      create_time:
        type: string
      created_by:
        type: string
      embedding_model:
        type: string
      id:
        type: string
      name:
        type: string
      update_time:
        type: string
      updated_by:
        type: string
    type: object
...
  request.KnowledgeBaseCreateRequest:
    properties:
      embedding_model:
        description: 嵌入模型标识
        example: qwen3-embedding:8b-fp16
        type: string
      name:
        description: 知识库名称
        example: 产品文档库
        type: string
    required:
    - name
    type: object
  request.KnowledgeBaseUpdateRequest:
    properties:
      name:
        description: 知识库名称
        example: 产品文档库（更新）
        type: string
    required:
    - name
    type: object
...
  response.Response:
    description: 统一响应格式
    properties:
      code:
        description: 状态码，0表示成功，非0表示失败
        example: 0
        type: integer
      data:
        description: 响应数据（可选）
      message:
        description: 响应消息
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Ragent Go 项目 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ragent Go API
  version: "1.0"
paths:
...
  /documents:
    get:
      description: 根据知识库ID分页查询文档列表
      parameters:
      - description: 知识库ID
        in: query
        name: kb_id
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询文档列表
      tags:
      - 文档
    post:
      consumes:
      - multipart/form-data
      description: 上传文档到知识库，自动进行解析、分块、向量化
      parameters:
      - description: 知识库ID
        in: formData
        name: kb_id
        required: true
        type: string
      - description: 文档文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Document'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 上传失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传文档
      tags:
      - 文档
  /documents/{id}:
    delete:
      description: 删除文档及其所有分块和向量数据
      parameters:
      - description: 文档ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文档不存在
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除文档
      tags:
      - 文档
    get:
      description: 根据ID获取文档详细信息
      parameters:
      - description: 文档ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Document'
              type: object
        "404":
          description: 文档不存在
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取文档详情
      tags:
      - 文档
  /documents/{id}/progress:
    get:
      description: 获取文档的处理进度信息（状态、分块数量、向量化进度等）
      parameters:
      - description: 文档ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "404":
          description: 文档不存在
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取文档处理进度
      tags:
      - 文档
...
  /knowledge-base:
    get:
      consumes:
      - application/json
      description: 分页查询知识库列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/model.KnowledgeBase'
                      type: array
                    page:
                      type: integer
                    page_size:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询知识库列表
      tags:
      - 知识库
    post:
      consumes:
      - application/json
      description: 创建新知识库，自动创建 Milvus Collection
      parameters:
      - description: 创建知识库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeBaseCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.KnowledgeBase'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建知识库
      tags:
      - 知识库
  /knowledge-base/{id}:
    delete:
      consumes:
      - application/json
      description: 删除知识库，同时删除 Milvus Collection
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除知识库
      tags:
      - 知识库
    get:
      consumes:
      - application/json
      description: 根据ID获取知识库详细信息
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.KnowledgeBase'
              type: object
        "404":
          description: 知识库不存在
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取知识库详情
      tags:
      - 知识库
    put:
      consumes:
      - application/json
      description: 更新知识库名称（只能更新名称）
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新知识库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeBaseUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新知识库
      tags:
      - 知识库
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer Token 进行认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"